Settlo/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
streamlit
pandas
python-multipart
google-generativeai
